// src/models/better-commerce/IPaymentInfo.ts
export interface IPaymentInfo {
    readonly paymentInfo1?: string | null
    readonly paymentInfo2?: string | null
    readonly paymentInfo3?: string | null
    readonly paymentInfo4?: string | null
    readonly paymentInfo5?: string | null
    readonly paymentInfo6?: string | null
    readonly paymentInfo7?: string | null
    readonly paymentInfo8?: string | null
}